Refactor middlewares and services of boilerplate server project
